evaluating the polynomial using Horner's method
Design and develop an algorithm for evaluating the polynomial f(x)  = a4x4 + a3x3 + a2x2 + a1x + a0,"
for a given value of x and its coefficients using Horner's method. Implement a Python program for the same and execute the program for different sets of values of coefficients and x.